fix into STDmcpy function (missing rts :() add macros to read a flow from memory add a way to emulate STe balance in PC emulation mode

+++ DEMOSDK/STANDARD.C
@@ -122,12 +122,12 @@ u16 STDifrnd()
 
 u32 STDmulu (u16 _x, u16 _y)
 {
-    return _x * _y;
+    return (u32)_x * (u32)_y;
 }
 
 s32 STDmuls (s16 _x, s16 _y)
 {
-    return _x * _y;
+    return (s32)_x * (s32)_y;
 }
 
 u32 STDdivu (u32 _x, u16 _y)
